pyproject.toml
[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[project]
name = "supply-disruption-env"
version = "1.0.0"
description = "OpenEnv supply chain disruption RL environment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "openenv-core",
    "fastapi",
    "uvicorn[standard]",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "server*", "agent*"]
